+++ TP3/src/sched.h
@@ -46,7 +46,7 @@ extern sched_t scheduler;
 void sched_agregar_tarea(perro_t *perro, uint gdt_index);
 
 // debe remover tareas del scheduler
-void sched_remover_tarea(unsigned int gdt_index);
+void sched_remover_tarea(unsigned short gdt_index);
 
 // debe devolver el perro correspondiente a la tarea que está corriendo actualmente
 perro_t* sched_tarea_actual();
